fix: tighten invoice triage integration anti-goals
