+++ examples/invoice-triage/src/bin.ts
@@ -1,0 +1,3 @@
+import { main } from "./main"
+
+await main()
