+++ examples/invoice-triage/src/main.ts
@@ -2,16 +2,18 @@ import { createScope } from "@pumped-fn/lite"
 import { logging, type Logging } from "@pumped-fn/lite-extension-logging"
 import { scheduler } from "@pumped-fn/lite-extension-scheduler"
 import { model as provider } from "@pumped-fn/sdk"
-import { pathToFileURL } from "node:url"
-import { OperationalFault } from "./errors"
-import { awaitDrained, dailyReportJob, ingest, intake, prepareDatabase, sendRemindersJob, watchReviewQueue } from "./flows"
-import type { DatabaseStartupMode } from "./migrations"
-import { databaseStartup, heuristic } from "./ports"
+import { dailyReportJob, ingest, prepareDatabase, sendRemindersJob, watchReviewQueue } from "./flows"
+import { databaseStartup, heuristic, mailer, type Mailer } from "./ports"
 
 export async function main(): Promise<void> {
   const sink: Logging.Sink = {
     name: "stdout",
     write: (record) => console.log(JSON.stringify(record)),
+  }
+  const reminders: Mailer = {
+    async send(message) {
+      console.log(JSON.stringify({ event: "invoice.reminder", message }))
+    },
   }
   const scope = createScope({
     extensions: [logging.extension()],
@@ -24,11 +26,10 @@ export async function main(): Promise<void> {
         fields: { service: "invoice-triage" },
       }),
       provider(heuristic),
-      databaseStartup(startupMode()),
+      databaseStartup("migrate"),
+      mailer(reminders),
     ],
   })
-  const stopIntake = () => process.stdin.destroy()
-  process.once("SIGINT", stopIntake)
 
   const ctx = scope.createContext()
   await ctx.exec({ flow: prepareDatabase })
@@ -36,21 +37,5 @@ export async function main(): Promise<void> {
   const watching = ctx.exec({ flow: watchReviewQueue })
   await ctx.resolve(dailyReportJob)
   await ctx.resolve(sendRemindersJob)
-  await ctx.exec({ flow: intake })
-  await ctx.exec({ flow: awaitDrained })
-  await ctx.close({ ok: true })
-  await scope.dispose()
-  await Promise.allSettled([processing, watching])
-  process.off("SIGINT", stopIntake)
-}
-
-if (import.meta.url === pathToFileURL(process.argv[1] ?? "").href) {
-  await main()
-}
-
-function startupMode(): DatabaseStartupMode {
-  const value = process.env["INVOICE_DATABASE_STARTUP"]
-  if (value === undefined || value === "" || value === "migrate") return "migrate"
-  if (value === "verify") return "verify"
-  throw new OperationalFault("unsupported-runtime-config", "read", "INVOICE_DATABASE_STARTUP", { value })
+  await Promise.all([processing, watching])
 }
