+++ examples/invoice-triage/src/ports.ts
@@ -1,5 +1,4 @@
 import { atom, tag, tags } from "@pumped-fn/lite"
-import { createInterface, type Interface } from "node:readline"
 import type { Model } from "@pumped-fn/sdk"
 import { databaseEngine } from "./database"
 import { pushFeed, type PushFeed } from "./feed"
@@ -13,7 +12,6 @@ export interface Clock {
 
 export interface Mailer {
   send(message: ReminderMessage): Promise<void>
-  sent(): readonly ReminderMessage[]
 }
 
 export interface OpsHeartbeat {
@@ -40,7 +38,7 @@ export const databaseStartup = tag<DatabaseStartupMode>({
   label: "invoice.databaseStartup",
 })
 
-export { databaseEngine, memoryDatabase, postgresDatabase } from "./database"
+export { databaseEngine, postgresDatabase } from "./database"
 
 export const database = atom({
   keepAlive: true,
@@ -54,12 +52,8 @@ export const database = atom({
   },
 })
 
-export const intakeLines = atom({
-  factory: (ctx): AsyncIterable<string> => {
-    const lines = createStdinLines()
-    ctx.cleanup(() => lines.close())
-    return lines
-  },
+export const intakeLines = tag<AsyncIterable<string>>({
+  label: "invoice.intakeLines",
 })
 
 export const opsHeartbeat = atom({
@@ -73,9 +67,8 @@ export const opsHeartbeat = atom({
   },
 })
 
-export const mailer = atom({
-  keepAlive: true,
-  factory: memoryMailer,
+export const mailer = tag<Mailer>({
+  label: "invoice.mailer",
 })
 
 export const heuristic: Model = Object.freeze({
@@ -90,19 +83,3 @@ export const heuristic: Model = Object.freeze({
     }
   },
 })
-
-function createStdinLines(): Interface {
-  return createInterface({ input: process.stdin })
-}
-
-export function memoryMailer(): Mailer {
-  const messages: ReminderMessage[] = []
-  return {
-    async send(message) {
-      messages.push(message)
-    },
-    sent() {
-      return messages.slice()
-    },
-  }
-}
