+++ examples/invoice-triage/tests/support/mailer.ts
@@ -1,0 +1,18 @@
+import type { Mailer } from "../../src/ports"
+import type { ReminderMessage } from "../../src/types"
+
+export interface MemoryMailer extends Mailer {
+  sent(): readonly ReminderMessage[]
+}
+
+export function memoryMailer(): MemoryMailer {
+  const messages: ReminderMessage[] = []
+  return {
+    async send(message) {
+      messages.push(message)
+    },
+    sent() {
+      return messages.slice()
+    },
+  }
+}
